+++ packages/visual-dev/src/utlis.ts
@@ -1,0 +1,4 @@
+import { StyleProps as BadgeStyleProps } from "./components/Badge";
+
+export const dashToSnakeCase = (str: BadgeStyleProps["status"]) =>
+  str.replace("-", "_");